requirements.txt
streamlit
PyPDF2
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests